requirement.txt
pytorch
pyyaml